Factors out PriorKnowledge and ContentCheck from NormalizationService Fixes a bug with linkedInPublicProfile regex matching
